[EventPipe] Adding an IPC server to handle out of process communication. (dotnet/coreclr#23106)
This is the first commit to enable a "diagnostic port" using IPC (Named Pipe on Windows and Unix Domain Socket on other platforms). This change currently enable EventPipe to be enabled/disabled without the use of a file drop.
- Split the DiagnosticsIpc into (DiagnosticsIpc/IpcStream)
  - DiagnosticsIpc (IPC listener) is meant to be used by the Diagnostic server.
  - IpcStream (IPC channel) is meant to be use to communicate with the connected client.
- Change the FastSerializer dependency from `CFileStream` to `StreamWriter`
  This abstraction is meant decouple the writing of objects in order to extend its usability.
  The main objective is to reuse FastSerializer to stream data through the open IPC channel.
- Moved the EventPipeSessionProvider* classes to their own file.
- Added a more streamlined parsing achievable by defining a simpler binary protocol (by noahfalk).
  1. Only one allocation is needed for the EventPipeProviderConfiguration array, no allocations or copies are needed for strings because we can refer to them directly out of the incoming command buffer
  2. No change to the EventPipe API for enable is required. EventPipeProviderConfiguration retains its current behavior of not deleting the string pointers it holds.
  3. No leaks happen because the command buffer owns the string memory and ensures that it stays alive for the duration of the Enable() call.

Commit migrated from https://github.com/dotnet/coreclr/commit/4497df34cbe3d8bf7a21828714a3cf764da16e98

+++ src/coreclr/src/System.Private.CoreLib/src/System/Diagnostics/Eventing/EventPipe.cs
@@ -196,9 +196,9 @@ namespace System.Diagnostics.Tracing
             s_sessionID = EventPipeInternal.Enable(
                 configuration.OutputFile,
                 configuration.CircularBufferSizeInMB,
-                configuration.ProfilerSamplingRateInNanoseconds,
+                (ulong)configuration.ProfilerSamplingRateInNanoseconds,
                 providers,
-                providers.Length,
+                (uint)providers.Length,
                 configuration.MultiFileTraceLengthInSeconds);
         }
 
@@ -214,7 +214,13 @@ namespace System.Diagnostics.Tracing
         // These PInvokes are used by the configuration APIs to interact with EventPipe.
         //
         [DllImport(JitHelpers.QCall, CharSet = CharSet.Unicode)]
-        internal static extern UInt64 Enable(string outputFile, uint circularBufferSizeInMB, long profilerSamplingRateInNanoseconds, EventPipeProviderConfiguration[] providers, int numProviders, ulong multiFileTraceLengthInSeconds);
+        internal static extern UInt64 Enable(
+            string outputFile,
+            uint circularBufferSizeInMB,
+            ulong profilerSamplingRateInNanoseconds,
+            EventPipeProviderConfiguration[] providers,
+            uint numProviders,
+            ulong multiFileTraceLengthInSeconds);
 
         [DllImport(JitHelpers.QCall, CharSet = CharSet.Unicode)]
         internal static extern void Disable(UInt64 sessionID);
